+++ ConquerorsCrown_NewHorizon/Motor2D/j1Scene.h
@@ -53,6 +53,7 @@ public:
 	p2SString current_level;
 	SDL_Texture* Hello;
 	SDL_Texture* debug_tex;
+	GuiItem* buttonNewGame;
 	bool debug;
 };
